update: ws多bot基础设施完成

+++ provider/bots_event/handler.go
@@ -6,7 +6,7 @@ import (
 )
 
 type handler struct {
-	BotAccount int
+	BotAccount int64
 }
 
 func (c *handler) OnOpen(socket *gws.Conn) {
